fix test and code duplication
