+++ webpack/src/prodsys/components/ModelField/EmailField.js
@@ -1,5 +1,6 @@
 // Email Field
 import { stringify } from 'utils/text'
+import { mailTo } from 'utils/phonemail'
 
 export const EditableField = ({ value, ...args }) => (
   <input
@@ -12,24 +13,3 @@ export const EditableField = ({ value, ...args }) => (
 export const DetailField = ({ value, ...args }) => (
   <span {...args}>{mailTo(value)}</span>
 )
-
-const phoneLink = R.pipe(
-  R.defaultTo(''),
-  R.replace(/\s/g, ''),
-  R.when(
-    R.pipe(
-      R.length,
-      R.equals(8),
-    ),
-    R.concat('+47'),
-  ),
-  R.concat('sms://'),
-)
-
-const phoneTo = R.ifElse(R.not, R.always('–'), phone => (
-  <a href={phoneLink(phone)}>{phoneFormat(phone)}</a>
-))
-
-const mailTo = R.ifElse(R.not, R.always('–'), mail => (
-  <a href={`mailto:${mail}`}>{R.replace(/@/, '\u200B@', mail)}</a>
-))
