+++ webpack/src/universitas/components/ContactCard/ContactCard.js
@@ -1,4 +1,5 @@
 import { capitalize, phoneFormat } from 'utils/text'
+import { phoneTo, mailTo } from 'utils/phonemail'
 import anonymous from 'images/anonymous.jpg'
 import cx from 'classnames'
 import './ContactCard.scss'
@@ -9,27 +10,6 @@ const Field = ({ name, label, value }) => (
     <div className="value">{value}</div>
   </div>
 )
-
-const phoneLink = R.pipe(
-  R.defaultTo(''),
-  R.replace(/\s/g, ''),
-  R.when(
-    R.pipe(
-      R.length,
-      R.equals(8),
-    ),
-    R.concat('+47'),
-  ),
-  R.concat('sms://'),
-)
-
-const phoneTo = R.ifElse(R.not, R.always('–'), phone => (
-  <a href={phoneLink(phone)}>{phoneFormat(phone)}</a>
-))
-
-const mailTo = R.ifElse(R.not, R.always('–'), mail => (
-  <a href={`mailto:${mail}`}>{R.replace(/@/, '\u200B@', mail)}</a>
-))
 
 export const ContactCard = ({
   id,
